* work on issue #16 (logical image size configurable at startup via cli arg)
 * default resolution is now 720p (rather than 1080p, makes program seem faster)

 * open-state-file now preselects the most recently saved state file

+++ src/ImageRegion.java
@@ -56,19 +56,19 @@ public class ImageRegion implements Serializable {
     }
 
     public int getXMin(){
-        return (int) upperLeft.getX();
+        return upperLeft.getX();
     }
 
     public int getYMin(){
-        return (int) upperLeft.getY();
+        return upperLeft.getY();
     }
 
     public int getXMax(){
-        return (int) lowerRight.getX();
+        return lowerRight.getX();
     }
 
     public int getYMax(){
-        return (int) lowerRight.getY();
+        return lowerRight.getY();
     }
 
     public int getWidth(){
